fix laser.last_firing casting

+++ Marlin/Marlin_main.cpp
@@ -2773,7 +2773,7 @@ void manage_inactivity()
   #endif
 
   #ifdef LASER
-  if (current_block->laser_duration > 0 && ((uint16_t)micros() - laser.last_firing) >= current_block->laser_duration) {
+  if (current_block->laser_duration > 0 && (micros() - laser.last_firing) >= current_block->laser_duration) {
     laser_extinguish();
   }  
   #endif // LASER
